kafka-4295: ConsoleConsumer does not delete the temporary group in zookeeper
Addressed Ijuma's comments
1. Restored ZkUtils to trunk code
2. Restored ConsoleConsumerTest to trunk code
3. Restored ZkUtils.maybeDeletePath to trunk code
4. Replaced ZkUtils.maybeDeletePath with
AdminUtils.deleteConsumerGroupInZK
